Update README.md
Added note about costs of keeping the server, constantly, up and running...in powered on/or, off mode...charges still accumulate. The best way is to delete the server altogether each time you use it...if you really wish to keep the costs as low as possible. Otherwise, use VMWare...and, download a Linux distro to run on that....where you can play around with learning Linux on Windows, and, all entirely for FREE.